+++ README.md
@@ -24,7 +24,7 @@ pip install -r requirements.txt
 In order to execute the script for the week starting on, e.g., April 9th, 2024, one needs to execute
 
 ```
-python3.11 main_script.py 2024 4 9
+python -m src.main_script 2024 4 9
 ```
 
 and a csv file in the `/out` folder is generated with the transfers.
@@ -32,6 +32,6 @@ and a csv file in the `/out` folder is generated with the transfers.
 There is also an optional argument that can be passed, which removes the transfers (not the accounting) of the rewards of the Gnosis solvers. E.g., usage
 
 ```
-python3.11 main_script.py 2024 4 9 ignore_gnosis_transfers
+python -m src.main_script 2024 4 9 ignore_gnosis_transfers
 ```
