+++ src/gnosis_scan_api.py
@@ -1,7 +1,7 @@
 import requests
 import datetime
 import calendar
-from constants import REQUEST_TIMEOUT, SUCCESS_CODE
+from src.constants import REQUEST_TIMEOUT, SUCCESS_CODE
 
 
 def get_block_range(year, month, day):
