+++ src/queries/auction_range.sql
