+++ src/queries/participation_rewards_aux.sql
