+++ src/queries/quote_rewards.sql
