+++ src/queries/solver_rewards.sql
