+++ src/queries/successful_auction_ids.sql
